+++ packages/core/src/base/renderer.ts
@@ -143,14 +143,10 @@ function patchComponent(nativeRenderer: NativeRenderer,
   const { computedContainer, contextContainer } = component.viewMetadata
   popContainer()
   if (portalContainer) {
-    if (portalContainer === context.contextContainer) {
-      if (portalContainer !== computedContainer) {
-        needMove = true
-      }
-    } else {
-      if (portalContainer !== computedContainer) {
-        needMove = true
-      }
+    if (portalContainer !== computedContainer) {
+      needMove = true
+    }
+    if (portalContainer !== context.contextContainer) {
       context = {
         isParent: true,
         anchorNode: portalContainer,
@@ -189,9 +185,12 @@ function deepUpdateByComponentDirtyTree(nativeRenderer: NativeRenderer, componen
     }
     component.rendered()
   } else if (component.changed) {
-    component.changedSubComponents.forEach(child => {
-      deepUpdateByComponentDirtyTree(nativeRenderer, child, needMove)
-    })
+    const changedSubComponents = component.changedSubComponents
+    if (changedSubComponents) {
+      for (const child of changedSubComponents) {
+        deepUpdateByComponentDirtyTree(nativeRenderer, child, needMove)
+      }
+    }
     component.rendered()
   }
 }
@@ -375,7 +374,7 @@ function updateComponent(nativeRenderer: NativeRenderer,
       ...context
     }
     newAtom.child = oldAtom.child
-    const skipSubComponentDiff = !component.changedSubComponents.size
+    const skipSubComponentDiff = !component.changedSubComponents?.size
     reuseComponentView(nativeRenderer, newAtom.child, context, needMove, skipSubComponentDiff)
     if (!skipSubComponentDiff) {
       component.rendered()
@@ -774,7 +773,6 @@ function updateNativeNodeProperties(
 
   if (!updatedChildren) {
     newAtom.child = oldAtom.child
-    // reuseElementChildrenView(nativeRenderer, newAtom, context)
   }
 
   if (bindRefs === unBindRefs) {